+++ src/generators/island.rs
@@ -1,11 +1,7 @@
-use std::sync::mpsc::Sender;
-
 use eframe::egui;
 use serde::{Deserialize, Serialize};
 
-use crate::ThreadMessage;
-
-use super::{get_min_max, report_progress};
+use super::{get_min_max, Progress};
 
 #[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
 pub struct IslandConf {
@@ -33,14 +29,11 @@ pub fn gen_island(
     size: (usize, usize),
     hmap: &mut [f32],
     conf: &IslandConf,
-    export: bool,
-    tx: Sender<ThreadMessage>,
-    min_progress_step: f32,
+    progress: &mut Progress,
 ) {
     let coast_h_dist = size.0 as f32 * conf.coast_range / 100.0;
     let coast_v_dist = size.1 as f32 * conf.coast_range / 100.0;
     let (min, _) = get_min_max(hmap);
-    let mut progress = 0.0;
     for x in 0..size.0 {
         for y in 0..coast_v_dist as usize {
             let h_coef = y as f32 / coast_v_dist as f32;
@@ -49,10 +42,8 @@ pub fn gen_island(
             let h = hmap[x + (size.1 - 1 - y) * size.0];
             hmap[x + (size.1 - 1 - y) * size.0] = (h - min) * h_coef + min;
         }
-        let new_progress = 0.5 * x as f32 / size.0 as f32;
-        if new_progress - progress >= min_progress_step {
-            progress = new_progress;
-            report_progress(progress, export, tx.clone());
+        if !progress.report(0.5 * x as f32 / size.0 as f32) {
+            return;
         }
     }
     for y in 0..size.1 {
@@ -63,10 +54,8 @@ pub fn gen_island(
             let h = hmap[(size.0 - 1 - x) + y * size.0];
             hmap[(size.0 - 1 - x) + y * size.0] = (h - min) * h_coef + min;
         }
-        let new_progress = 0.5 + 0.5 * y as f32 / size.0 as f32;
-        if new_progress - progress >= min_progress_step {
-            progress = new_progress;
-            report_progress(progress, export, tx.clone());
+        if !progress.report(0.5 + 0.5 * y as f32 / size.1 as f32) {
+            return;
         }
     }
 }